Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@5179 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/items/attachment.cpp
@@ -189,7 +189,7 @@ void Attachment::update(float dt)
     case ATTACH_BOMB:
         if(m_time_left <= (m_node->getEndFrame() - m_node->getStartFrame() - 1))
         {
-            m_node->setCurrentFrame((m_node->getEndFrame() - m_node->getStartFrame() - 1) - (int) m_time_left);
+            m_node->setCurrentFrame(m_node->getEndFrame() - m_node->getStartFrame() - 1 -m_time_left);
         }
         if(m_time_left<=0.0)
         {
